net: dsa: realtek: remove deprecated custom slave mii
Drop the code path using the custom slave mii. It still warns if there
is a compatible string in "mdio" node. However, if that compatible
string is in a node not named "mdio", it will show an error. In that
case, it might still work using polling instead of interruptions.

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>