tagger: update format with EFID_EN/EFID, VSEL/VIDX; set KEEP=1
We'll be able to use EFID to implement tx forward offload, but let's not
do it now.

KEEP=1 is needed because we should respect the format of the packet as
the kernel sends it to us. Unless tx forward offloading is used, the
kernel is giving us the packet exactly as it should leave the specified
port on the wire. Until now this was not needed because the ports were
always functioning in a standalone mode in a VLAN-unaware way, so the
switch would not tag or untag frames anyway. But arguably it should have
been KEEP=1 all along.