net: dsa: realtek: load switch variants on demand
realtek-common had a hard dependency on both switch variants. As a
result, it was not possible to selectively load only one model at
runtime. Now, variants are registered in the realtek-common module, and
interface modules look for a variant using the compatible string.

The variant modules use the same compatible string as the module alias.
This way, an interface module can use the matching compatible string to
both load the module and get the variant reference.

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>