net: dsa: realtek: rtl8365mb: add SGMII and HSGMII mode support
This adds support for SGMII and HSGMII on RTL8367S switches. HSGMII is
configured using the 2500BASEX mode. This is based on the rtl8367c
driver distributed by the vendor and located in GPL tarballs.

For (H)SGMII mode we have to load a firmware into some memory which gets
executed on the integrated 8051. The firmware is expected to be at
/lib/firmware/realtek/rtl8365mb/rtl8365mb-dw8051-sgmii-init.bin

This was tested on RTL8367S (ver=0x00a0, opt=0x0001).

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Schneider Azima <Schneider-Azima12@protonmail.com>
Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>